tests/requirements.txt
pytest==7.3.1
pytest-asyncio==0.21.0
pytest-cov==4.1.0
requests==2.31.0
httpx==0.24.1
pillow==9.5.0
numpy==1.24.3
opencv-python==4.7.0.72
aiofiles==23.1.0
scikit-learn==1.2.2
pytest-mock==3.10.0
pytest-timeout==2.1.0
pytest-docker==1.0.1
tenacity==8.2.2
fastapi==0.95.2
pydantic==1.10.8
aiohttp==3.8.5
# Below are additional dependencies for API tests
qdrant-client>=1.4.0
mysql-connector-python>=8.0.0
openai>=1.0.0
azure-identity>=1.14.0
azure-keyvault-secrets>=4.7.0 
